+++ src/components/AutoSuggestion/index.ts
@@ -1,7 +1,7 @@
-import { AutoSuggestion, AutoSuggestionProps } from "./AutoSuggestion";
+import { AutoSuggestionProps } from "./AutoSuggestion";
 import { AutoSuggestionList, AutoSuggestionListProps } from "./AutoSuggestionList";
 import { ExtendedCodeEditor, ExtendedCodeEditorProps } from "./ExtendedCodeEditor";
 
-export { AutoSuggestion, AutoSuggestionList, ExtendedCodeEditor };
+export { AutoSuggestionList, ExtendedCodeEditor };
 
 export type { AutoSuggestionProps, AutoSuggestionListProps, ExtendedCodeEditorProps };