+++ src/components/AutocompleteField/interfaces.ts
@@ -1,4 +1,4 @@
-export interface IRenderModifiers {
+export interface SuggestFieldItemRendererModifierProps {
     active: boolean;
     disabled?: boolean;
     // The width styles that should be given to the rendered option items
@@ -6,7 +6,10 @@ export interface IRenderModifiers {
     highlightingEnabled: boolean;
 }
 
-/** Style object to be used in menu option items. */
+/** @deprecated (v25) use `SuggestFieldItemRendererModifierProps` */
+export type IRenderModifiers = SuggestFieldItemRendererModifierProps;
+
+/** @deprecated (v25) use `SuggestFieldItemRendererModifierProps["styleWidth"]` */
 export interface IElementWidth {
     minWidth: string;
     maxWidth: string;