+++ CHANGELOG.md
@@ -37,7 +37,61 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/) and this p
 -   upgrade to Storybook v8
     -   include a few patches for actions, see https://github.com/storybookjs/storybook/blob/next/MIGRATION.md#implicit-actions-can-not-be-used-during-rendering-for-example-in-the-play-function
 -   allow `next` and `legacy` as branch names
+-   CodeMirror `setInstance` interface changed to `setEditorView` for semantic compatibility to version 6
 -   switch icons for `item-clone` and `item-copy` to Carbon's `<Replicate/>` and `<Copy/>`
+-   Added new properties to `<CodeMirror>`
+    -   `supportCodeFolding` optional property to fold code for the supported modes e.g: xml, json etc.
+    -   `shouldHighlightActiveLine` optional property to highlight active line where the cursor is currently in.
+    -   `shouldHaveMinimalSetup` optional property that imports codemirror's base minimal configurations.
+    -   `additionalExtensions` optional property for additional extensions to customize the editor further.
+
+### Deprecated
+
+-   `<Icon/>` and `<TestIcon/>`
+    -   `description` and `iconTitle`: use `title` as replacement.
+-   `TableRowHeightSize` type: use `TableProps["size"]` directly
+-   `IRenderModifiers` interface: use `SuggestFieldItemRendererModifierProps`
+-   `IElementWidth` type: use `SuggestFieldItemRendererModifierProps["styleWidth"]`
+-   `MultiSelectSelectionProps` interface: use `MultiSuggestFieldSelectionProps`
+-   `MultiSelectProps` interface: use `MultiSuggestFieldProps`
+-   `nodeTypes` and `edgeTypes`
+    -   will be removed without replacement, define it yourself or use `<ReactFlow/` with `configuration` option
+-   `AutoCompleteFieldProps` and `IAutoCompleteFieldProps` interfaces: use `SuggestFieldProps`
+-   `<CodeAutocompleteField/>`
+    -   `AutoSuggestionProps`: use `CodeAutocompleteFieldProps` instead
+    -   we renamed `ISuggestionBase`, `ISuggestionWithReplacementInfo`, `IReplacementResult`, `IPartialAutoCompleteResult`, `IValidationResult` to `CodeAutocompleteFieldSuggestionBase`, `CodeAutocompleteFieldSuggestionWithReplacementInfo`, `CodeAutocompleteFieldReplacementResult`, `CodeAutocompleteFieldPartialAutoCompleteResult`, `CodeAutocompleteFieldValidationResult`
+-   all legacy support components are going to be removed, you need to replace them by activily maintained components
+    -   `<ButtonReplacement/>`: switch to `<Button />`
+    -   `<AffirmativeButtonReplacement/>`: switch to `<Button affirmative />`
+    -   `<DismissiveButtonReplacement/>`: switch to `<Button dismissive />`
+    -   `<DisruptiveButtonReplacement/>`: switch to `<Button disruptive />`
+    -   `<CheckboxReplacement/>`: switch to `<Checkbox />`
+    -   `<RadioButtonReplacement/>`: switch to `<RadioButton />`
+    -   `<TabsReplacement/>`: switch to `<Tabs />`
+    -   `<TextFieldReplacement/>`: switch to `<TextField />`, `<TextArea />`, `<FieldItem />`
+-   `MultiSuggestField.ofType` method:
+    -   instead of `MyMultiSuggest = MultiSuggestField.ofType<MyType>()` use directly `<MultiSuggestField<MyType> {...props} />`
+    -   `MultiSuggestField.ofType` also returns the original BlueprintJS `MultiSelect` element, not our version!
+
+### Migration from v23 to v24
+
+-   upgrade Typescript to v5
+-   upgrade Node to at least v18, see **Changed** section for more info about it
+-   remove deprecated components, properties and imports from your project, if the info cannot be found here then it was already mentioned in **Deprecated** sections of the past changelogs
+    -   `<GridColumn/>`
+        -   `full`: was deprecated and now removed because it always uses full width if it is the only column and does not have any othe size config
+    -   `<Notification/>`
+        -   `fullWidth`: was deprecated and now removed, use `flexWidth` as replacement
+        -   `iconName`: was deprecated and now removed, use `icon` property
+    -   `<Table/>`
+        -   `size`: use only "small", "medium" or "large" as value
+    -   `<Tag/>`
+        -   `emphasized`: was deprecated and now removed, use `minimal=false` plus `emphasis="stronger"` instead
+    -   `IconSized` type: use `CarbonIconType`
+    -   `TimeUnits` type: use `ElapsedDateTimeDisplayUnits`
+    -   `MarkdownParserProps` interface: use `MarkdownProps`
+    -   `elapsedTimeSegmented` function: use `elapsedDateTimeDisplayUtils.elapsedTimeSegmented`
+    -   `simplifiedElapsedTime` function: use `elapsedDateTimeDisplayUtils.simplifiedElapsedTime`
 
 ## [23.8.0] - 2024-08-19
 
@@ -86,7 +140,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/) and this p
     -   Element wraps the content that need to be displayed sticky.
 -   `utils`
     -   `getScrollParent`: method to find the scroll parent of an element
--   `<AutoCompleteField />`
+-   `<SuggestField />`
     -   Support loading more results when scrolling to the end of the result list.
 -   `<TextArea />`
     -   `intent` property to set the state, formerly used `hasStatePrimary`, `hasStateSuccess`, `hasStateWarning` and `hasStateDanger` properties are now deprecated
@@ -221,7 +275,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/) and this p
 -   `<ActivityControlWidget />`
     -   added extra line to show timer for execution period
 -   `<ExtendedCodeEditor />`
-    -   replaces `<SingleLineCodeEditor />` to get used for the `<AutoSuggestion />` component
+    -   replaces `<SingleLineCodeEditor />` to get used for the `<CodeAutocompleteField />` component
 -   new icons
     -   `data-string`, `data-url`, `data-date`, `data-time`, `data-datetime`, `data-number`
 
@@ -269,7 +323,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/) and this p
     -   `loose`: can be set to `true` to prevent the box with border on the label component
 -   `<TableExpandHeader />`
     -   `toggleIcon`: optional icon that should be displayed instead of the default ones.
--   `Utilities`
+-   `utils`
     -   `getGlobalVar` and `setGlobalVar`: can be used to manage global variables indepentently from component states. They are stored to the `window` object under a `eccgui` "namespace". Can be used for example to manage globally increased counters. Do not use them if you need to store user session properties or confidential data!
 -   canonical icons for `artefact-chatlog`, `entity-human`, `entity-robot` and `operation-magic`
 
@@ -399,7 +453,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/) and this p
 -   `<WorkspaceHeader />`
     -   `IWorkspaceHeaderProps` interface is now deprecated, use `WorkspaceHeaderProps` instead
 -   `<NumericInput />`
-    -   It will be remove because beside the special arrow buttons it does not add any special. Could be done also with `<TextField />` combined with correct `type`.
+    -   It will be removed because beside the special arrow buttons it does not add any special. Could be done also with `<TextField />` combined with correct `type`.
 -   `<Highlighter />`
     -   `HighlighterFunctions` renamed to `highlighterUtils`
     -   `extractSearchWords` moved to `highlighterUtils.extractSearchWords`
@@ -414,7 +468,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/) and this p
 -   `ReactFlow` extensions
     -   `NodeProps`: renamed to `NodeDefaultProps`
     -   `minimapNodeClassName`: moved to `miniMapUtils.nodeClassName`
-    -   `minimapNodeColor`: moved to `miniMapUtils.nodeClassName`
+    -   `minimapNodeColor`: moved to `miniMapUtils.nodeColor`
     -   `nodeUtils`: renamed to `nodeDefaultUtils`
     -   `IHandleProps`: renamed to `NodeContentHandleProps`
     -   `NodeDimensions`: use `NodeContentProps<any>['nodeDimensions']`